Port request module and add extension id
